Items start (NOT COMPLETE)

+++ MortalGamblingProj/Assets/Scripts/Managers/TeamManager.cs
@@ -69,12 +69,12 @@ public class TeamManager : MonoBehaviour
         return _teams[teamId].GetPlayerAmount();
     }
 
-    public void ApplyTeamHealthChange(int teamIdx, int playerIdx, int healthChange)
+    public void ApplyTeamHealthChange(int teamIdx, int playerIdx, float healthChange)
     {
         _teams[teamIdx].ApplyHealthChange(healthChange, playerIdx);
     }
 
-    public void ApplyTeamStaminaChange(int teamIdx, int playerIdx, int staminaChange)
+    public void ApplyTeamStaminaChange(int teamIdx, int playerIdx, float staminaChange)
     {
         _teams[teamIdx].ApplyStaminaChange(staminaChange, playerIdx);
     }
